# server/swagger.yaml
openapi: 3.0.0
info:
  title: MacArea API documentation
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Local development server
paths:
...
  /scripts/create:
   post:
      tags:
      - "Scripts"
      summary: Create a new script
      description: Create a new script.
      operationId: createScript
      security:
        - bearerAuth: []
      parameters:
        - in: cookie
          name: refreshToken
          description: The refresh token can be sent in a cookie or the post body, if both are sent the token in the body is used.
          schema:
            type: string
            example: 51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                action_id:
                  type: string
                  example: "6040dac6678002004a1bafdd"
                reaction_id:
                  type: string
                  example: "6040dac6678002004a1bafe1"
                action_parameters:
                  type: array
                  items:
                    type: string
                    example: "john@mail.com"
                reaction_parameters:
                  type: array
                  items:
                    type: string
                    example: "message"
                trigger:
                  type: number
                  example: 10
      responses:
        "200":
          description: The email was successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Email successfully updated"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
  /scripts:
   get:
      tags:
      - "Scripts"
      summary: Get scripts
      description: Cet all user scripts.
      operationId: getAllScrpits
      security:
        - bearerAuth: []
      parameters:
        - in: cookie
          name: refreshToken
          description: The refresh token can be sent in a cookie or the post body, if both are sent the token in the body is used.
          schema:
            type: string
            example: 51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550
      responses:
        "200":
          description: The email was successfully updated
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/script"
  /scripts/{id}/update:
   post:
      tags:
      - "Scripts"
      summary: Update scripts status
      description: Update a script status
      operationId: updateScrpits
      security:
        - bearerAuth: []
      parameters:
        - in: cookie
          name: refreshToken
          description: The refresh token can be sent in a cookie or the post body, if both are sent the token in the body is used.
          schema:
            type: string
            example: 51872eca5efedcf424db4cf5afd16a9d00ad25b743a034c9c221afc85d18dcd5e4ad6e3f08607550
        - in: path
          name: id
          description: User id
          required: true
          example: "5eb12e197e06a76ccdefc121"
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  example: "start"
      responses:
        "200":
          description: The email was successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
components:
  schemas:
    script:
      type: object
      properties:
        id:
          type: string
          example: "60437c3e5222f70070749d3a"
        action:
          type: object
          properties:
            name:
              type: string
              example: "mail"
            description:
              type: string
              example: "New mail received"
        reaction:
          type: object
          properties:
            name:
              type: string
              example: "mail"
            description:
              type: string
              example: "Send a mail"
        trigger:
          type: number
          example: 10
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Access token is missing or invalid, or the user does not have access to perform the action
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Unauthorized"
...